+++ api/src/controllers/mail.js
@@ -8,7 +8,6 @@ const { capture } = require("../third-parties/sentry");
 router.post(
   "/",
   catchErrors(async (req, res) => {
-    console.log(JSON.stringify(req.body));
     let { to, replyTo, replyToName, subject, text, html } = req.body || {};
 
     if (!subject || (!text && !html)) return res.status(400).json({ ok: false, error: "wrong parameters" });